chore: cleaning file structure for button components
