--- packages/react/src/components/Button/types.ts
@@ -1,3 +1,0 @@
-export type ButtonSize = 'medium' | 'small';
-
-export type ButtonVariant = 'brand' | 'danger' | 'primary' | 'secondary' | 'tertiary';
