--- packages/react/src/components/ButtonGroup/ButtonGroup.context.ts
@@ -1,12 +1,0 @@
-import { createContext } from '../../utils/context';
-import type { ButtonSize, ButtonVariant } from '../Button/types';
-
-export interface ButtonGroupContext {
-  isDisabled?: boolean;
-  size?: ButtonSize;
-  variant?: ButtonVariant;
-}
-
-export const [ButtonGroupProvider, useButtonGroup] = createContext<ButtonGroupContext>({
-  name: 'ButtonGroupContext',
-});
