--- packages/react/src/components/ButtonGroup/index.ts
@@ -1,3 +1,0 @@
-export { ButtonGroup } from './ButtonGroup';
-export { useButtonGroup } from './ButtonGroup.context';
-export type { ButtonGroupElement, ButtonGroupProps } from './ButtonGroup.types';
