+++ packages/react/src/components/CalendarHeader/CalendarHeader.tsx
@@ -2,8 +2,8 @@ import { useDateFormatter, useLocale } from '@react-aria/i18n';
 import type { CalendarState, RangeCalendarState } from '@react-stately/calendar';
 import type { AriaButtonProps } from '@react-types/button';
 import { cx } from '@project44-manifest/react-styles';
+import { IconButton } from '../button';
 import { Icon } from '../Icon';
-import { IconButton } from '../IconButton';
 import { Typography } from '../Typography';
 import { useStyles } from './CalendarHeader.styles';
