--- packages/react/src/components/IconButton/IconButton.types.ts
@@ -1,5 +1,0 @@
-import type { ButtonElement, ButtonProps } from '../Button';
-
-export type IconButtonElement = ButtonElement;
-
-export type IconButtonProps = Omit<ButtonProps, 'endIcon' | 'startIcon'>;
