--- packages/react/src/components/IconButton/index.ts
@@ -1,2 +1,0 @@
-export { IconButton } from './IconButton';
-export type { IconButtonElement, IconButtonProps } from './IconButton.types';
