+++ packages/react/src/components/SideNavigationFooter/SideNavigationFooter.types.ts
@@ -1,6 +1,6 @@
 import type { CSS } from '@project44-manifest/react-styles';
 import type { AvatarProps } from '../avatar';
-import type { ButtonProps } from '../Button';
+import type { ButtonProps } from '../button';
 
 export type SideNavigationFooterElement = 'button';
