+++ packages/react/src/components/SideNavigationFooterMenu/SideNavigationFooterMenu.tsx
@@ -4,7 +4,7 @@ import { cx } from '@project44-manifest/react-styles';
 import type { ForwardRefComponent } from '@project44-manifest/react-types';
 import { mergeProps } from '../../utils';
 import { Avatar } from '../avatar';
-import { useButton, useHover } from '../Button';
+import { useButton, useHover } from '../button';
 import { Dropdown } from '../Dropdown';
 import { DropdownMenu, DropdownMenuProps } from '../DropdownMenu';
 import { Typography } from '../Typography';
