+++ packages/react/src/components/SideNavigationHeader/SideNavigationHeader.styles.ts
@@ -1,5 +1,5 @@
 import { pxToRem, styled } from '@project44-manifest/react-styles';
-import { IconButton } from '../IconButton';
+import { IconButton } from '../button';
 
 export const StyledSideNavigationHeaderLogo = styled('div', {
   alignItems: 'center',
