+++ packages/react/src/components/button/index.ts
@@ -1,5 +1,5 @@
-export { Button } from './Button';
-export type { ButtonElement, ButtonProps } from './Button.types';
+export * from './button';
 export { useButton } from '@react-aria/button';
 export { FocusRing, FocusScope, useFocusRing } from '@react-aria/focus';
 export { useFocus, useHover } from '@react-aria/interactions';
+export type { FocusableProps, PressEvents } from '@react-types/shared';
