+++ packages/react/src/index.ts
@@ -2,10 +2,7 @@ export * from './components/action-card';
 export * from './components/avatar';
 export * from './components/box';
 export * from './components/breadcrumb';
-export type { ButtonProps } from './components/Button';
-export { Button } from './components/Button';
-export type { ButtonGroupProps } from './components/ButtonGroup';
-export { ButtonGroup } from './components/ButtonGroup';
+export * from './components/button';
 export type { CalendarProps } from './components/Calendar';
 export { Calendar } from './components/Calendar';
 export type { CalendarRangeProps } from './components/CalendarRange';
@@ -57,8 +54,6 @@ export { FormControl } from './components/FormControl';
 export * from './components/grid';
 export type { IconProps } from './components/Icon';
 export { Icon } from './components/Icon';
-export type { IconButtonElement, IconButtonProps } from './components/IconButton';
-export { IconButton } from './components/IconButton';
 export type { LinkProps } from './components/Link';
 export { Link } from './components/Link';
 export type { ListBoxProps } from './components/ListBox';
